materials-ai/backend/requirements.txt
# FastAPI & Web
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
websockets>=12.0
python-multipart>=0.0.6
httpx>=0.26.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
pgvector>=0.2.4
alembic>=1.13.0

# Redis
redis>=5.0.0

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ML / AI
torch>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0

# Chemistry
rdkit-pypi>=2023.9.4

# Data processing
pandas>=2.1.0
pyarrow>=14.0.0
scipy>=1.12.0

# Optional: Uni-Mol (install separately if GPU available)
# unimol-tools>=1.0.0

# Optional: Additional ML
# torch-geometric>=2.4.0
# transformers>=4.36.0
# chemprop>=1.6.0
# deepchem>=2.7.0
# mordred>=1.2.0

# Utilities
python-dotenv>=1.0.0
